Added a feature to set the number of columns to 1
It seems we can restart the frame transfer by sending a FRAME signal. So we are going to try just sending 1 column, the DAC values. The selection bit is in WireIn10, bit 5. Was a bit of a fight as it could already be written to using the full register structure.  This should allow ramping up the DAC values to avoid high current, if this was the problem.